move test deps to codemodder deps
