+++ pyproject.toml
@@ -11,6 +11,7 @@ readme = "README.md"
 license = {file = "LICENSE"}
 description = "A pluggable framework for building codemods in Python"
 dependencies = [
+    "GitPython<4",
     "isort>=5.12,<5.14",
     "libcst>=1.1,<1.3",
     "packaging~=23.2",
@@ -48,7 +49,6 @@ get-hashes = 'codemodder.scripts.get_hashes:main'
 test = [
     "coverage>=7.3,<7.5",
     "Flask<4",
-    "GitPython<4",
     "Jinja2~=3.1.2",
     "lxml>=4.9.3,<5.2.0",
     "mock==5.1.*",
